+++ src/commands/command/inputs/index.ts
@@ -1,4 +1,6 @@
+// TODO: this should export everything that is visible outside this module
 export type { CreateTableInput } from "./create-table.js";
 export type { DeleteItemInput, DeleteItemReturnValue } from "./delete-item.js";
 export type { DeleteTableInput } from "./delete-table.js";
 export type { GetItemInput } from "./get-item.js";
+export type { PutItemInput } from "./put-item.js";