+++ src/commands/command/outputs/index.ts
@@ -2,3 +2,4 @@ export type { CreateTableOutput } from "./create-table.js";
 export type { DeleteItemOutput } from "./delete-item.js";
 export type { DeleteTableOutput } from "./delete-table.js";
 export type { GetItemOutput } from "./get-item.js";
+export type { PutItemOutput } from "./put-item.js";