+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.57.0] - 2026-02-19
+
+### Added
+
+- The `PutItem`command.
+
+### Changed
+
+- Renamed `PutItemParams` to `PutItemInput`, and `PutItemResult` to `PutItemOutput`.
+
+
 ## [0.56.0] - 2026-02-19
 
 ### Added
@@ -722,6 +733,7 @@ intuitive.
 - Initial release of the package! Move the implementation work in progress from another
 project to here.
 
+[0.57.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.56.0...v0.57.0
 [0.56.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.55.0...v0.56.0
 [0.55.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.54.0...v0.55.0
 [0.54.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.53.0...v0.54.0